fix bugs with order deletion

+++ src/service/OrderService.hx
@@ -1,5 +1,6 @@
 package service;
 import Common;
+import tink.core.Error;
 
 /**
  * Order Service
@@ -7,31 +8,62 @@ import Common;
  */
 class OrderService
 {
-	
+	/**
+	 *  Delete an order
+	 */
 	public static function delete(order:db.UserContract) {
-
 		var t = sugoi.i18n.Locale.texts;
+
+		if(order==null) throw new Error(t._("This order has already been deleted."));
+		
+		order.lock();
+		
 		if (order.quantity == 0) {
+
+			var contract = order.product.contract;
 			var user = order.user;
-			var place = order.distribution.place;
-			order.lock();
-			order.delete();
 
-			//Get the basket for this user
-			var basket = db.Basket.get(user, place, order.distribution.date);
+			//Amap Contract
+			if ( contract.type == db.Contract.TYPE_CONSTORDERS ) {
+				
+				order.delete();
 
-			//Get all the orders for this basket
-			var orders = basket.getOrders();
+				if( contract.amap.hasPayments() )
+				{
+					var orders = contract.getUserOrders(user);
+					if( orders.length == 0 ){
+						var operation = db.Operation.findCOrderTransactionFor(contract, user);
+						if(operation!=null) operation.delete();
+					}
+				}
 
-			//Check there is no orders left to delete the related operation
-			if( orders.length == 0 && place.amap.hasPayments())
-			{
-				var operation = db.Operation.findVOrderTransactionFor(order.distribution.getKey(), user, place.amap);
-				operation.delete();
+			}
+			else { //Variable orders contract
+
+				order.delete();
+
+				if( contract.amap.hasPayments() )
+				{
+
+					//Get the basket for this user
+					var place = order.distribution.place;
+					var basket = db.Basket.get(user, place, order.distribution.date);
+
+					//Get all the orders for this basket
+					var orders = basket.getOrders();
+
+					//Check there is no orders left to delete the related operation
+					if( orders.length == 0 )
+					{
+						var operation = db.Operation.findVOrderTransactionFor(order.distribution.getKey(), user, place.amap);
+						if(operation!=null) operation.delete();
+					}
+
+				}
 			}
 		}
 		else {
-			throw new tink.core.Error(t._("Deletion non possible: quantity is not zero."));
+			throw new Error(t._("Deletion not possible: quantity is not zero."));
 		}
 
 	}
